+++ sg_compute_specs/podman/schemas/Schema__Podman__List.py
@@ -1,0 +1,15 @@
+# ═══════════════════════════════════════════════════════════════════════════════
+# SG/Compute Specs — Podman: Schema__Podman__List
+# ═══════════════════════════════════════════════════════════════════════════════
+
+from osbot_utils.type_safe.Type_Safe                                                import Type_Safe
+
+from sgraph_ai_service_playwright__cli.observability.primitives.Safe_Str__AWS__Region import Safe_Str__AWS__Region
+
+from sg_compute_specs.podman.collections.List__Schema__Podman__Info                 import List__Schema__Podman__Info
+
+
+class Schema__Podman__List(Type_Safe):
+    region       : Safe_Str__AWS__Region
+    stacks       : List__Schema__Podman__Info
+    total        : int = 0